+++ addOns/authhelper/CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
+### Added
+- Support for handling the Microsoft login "Permissions requested" user consent screen.
 
 ### Fixed
 - Handling of Ionic input elements, which can appear to not be displayed until they are clicked on.